Fixed issue with previous commit not acting as attributes Addes Geo, Secondary, and Secondary Count datasets
