+++ smarty-rust-sdk/examples/us_enrichment_api.rs
@@ -1,11 +1,16 @@
 use smarty_rust_sdk::sdk::authentication::SecretKeyCredential;
 use smarty_rust_sdk::sdk::options::OptionsBuilder;
-use smarty_rust_sdk::us_enrichment_api::client::USEnrichmentClient;
+
+use smarty_rust_sdk::us_enrichment_api::client::*;
+use smarty_rust_sdk::us_enrichment_api::financial::*;
+use smarty_rust_sdk::us_enrichment_api::geo::*;
+use smarty_rust_sdk::us_enrichment_api::principal::*;
+
 use smarty_rust_sdk::us_enrichment_api::lookup::EnrichmentLookup;
-use smarty_rust_sdk::us_enrichment_api::results::{
-    EnrichmentResponse, FinancialResponse, GeoReference2010Response, GeoReference2020Response,
-    GeoReferenceResponse, PrincipalResponse,
-};
+use smarty_rust_sdk::us_enrichment_api::response::EnrichmentResponse;
+use smarty_rust_sdk::us_enrichment_api::secondary::SecondaryCountResponse;
+use smarty_rust_sdk::us_enrichment_api::secondary::SecondaryResponse;
+
 use std::error::Error;
 
 #[tokio::main]
@@ -17,6 +22,8 @@ async fn main() -> Result<(), Box<dyn Error>> {
     lookup::<GeoReferenceResponse>(key).await?;
     lookup::<GeoReference2010Response>(key).await?;
     lookup::<GeoReference2020Response>(key).await?;
+    lookup::<SecondaryResponse>(key).await?;
+    lookup::<SecondaryCountResponse>(key).await?;
 
     Ok(())
 }
