+++ smarty-rust-sdk/src/us_enrichment_api/mod.rs
@@ -1,6 +1,12 @@
 pub mod client;
 pub mod lookup;
-pub mod results;
+
+pub mod response;
+
+pub mod financial;
+pub mod geo;
+pub mod principal;
+pub mod secondary;
 
 #[cfg(test)]
 mod tests {
