+++ smarty-rust-sdk/src/us_enrichment_api/principal.rs
@@ -1,13 +1,8 @@
-use serde::de::DeserializeOwned;
 use serde::Deserialize;
 use serde::Serialize;
 use std::fmt::Debug;
-use std::ops::Deref;
-use std::ops::DerefMut;
 
-pub trait EnrichmentResponse: Clone + Serialize + DeserializeOwned + Default {
-    fn lookup_type() -> &'static str;
-}
+use super::response::EnrichmentResponse;
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct PrincipalResponse {
@@ -388,290 +383,4 @@ pub struct PrincipalAttributes {
     pub wine_cellar: String,
     pub year_built: String,
     pub zoning: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(default)]
-pub struct FinancialAttributes {
-    pub assessed_improvement_percent: String,
-    pub assessed_improvement_value: String,
-    pub assessed_land_value: String,
-    pub assessed_value: String,
-    pub assessor_last_update: String,
-    pub assessor_taxroll_update: String,
-    pub contact_city: String,
-    pub contact_crrt: String,
-    pub contact_full_address: String,
-    pub contact_house_number: String,
-    pub contact_mail_info_format: String,
-    pub contact_mail_info_privacy: String,
-    pub contact_mailing_county: String,
-    pub contact_mailing_fips: String,
-    pub contact_post_direction: String,
-    pub contact_pre_direction: String,
-    pub contact_state: String,
-    pub contact_street_name: String,
-    pub contact_suffix: String,
-    pub contact_unit_designator: String,
-    pub contact_value: String,
-    pub contact_zip: String,
-    pub contact_zip4: String,
-    pub deed_document_page: String,
-    pub deed_document_book: String,
-    pub deed_document_number: String,
-    pub deed_owner_first_name: String,
-    pub deed_owner_first_name2: String,
-    pub deed_owner_first_name3: String,
-    pub deed_owner_first_name4: String,
-    pub deed_owner_full_name: String,
-    pub deed_owner_full_name2: String,
-    pub deed_owner_full_name3: String,
-    pub deed_owner_full_name4: String,
-    pub deed_owner_last_name: String,
-    pub deed_owner_last_name2: String,
-    pub deed_owner_last_name3: String,
-    pub deed_owner_last_name4: String,
-    pub deed_owner_middle_name: String,
-    pub deed_owner_middle_name2: String,
-    pub deed_owner_middle_name3: String,
-    pub deed_owner_middle_name4: String,
-    pub deed_owner_suffix: String,
-    pub deed_owner_suffix2: String,
-    pub deed_owner_suffix3: String,
-    pub deed_owner_suffix4: String,
-    pub deed_sale_date: String,
-    pub deed_sale_price: String,
-    pub deed_transaction_id: String,
-    pub disabled_tax_exemption: String,
-    pub financial_history: Vec<FinancialHistory>,
-    pub first_name: String,
-    pub first_name_2: String,
-    pub first_name_3: String,
-    pub first_name_4: String,
-    pub homeowner_tax_exemption: String,
-    pub last_name: String,
-    pub last_name_2: String,
-    pub last_name_3: String,
-    pub last_name_4: String,
-    pub market_improvement_percent: String,
-    pub market_improvement_value: String,
-    pub market_land_value: String,
-    pub market_value_year: String,
-    pub match_type: String,
-    pub middle_name: String,
-    pub middle_name_2: String,
-    pub middle_name_3: String,
-    pub middle_name_4: String,
-    pub other_tax_exemption: String,
-    pub owner_full_name: String,
-    pub owner_full_name_2: String,
-    pub owner_full_name_3: String,
-    pub owner_full_name_4: String,
-    pub ownership_transfer_date: String,
-    pub ownership_transfer_doc_number: String,
-    pub ownership_transfer_transaction_id: String,
-    pub ownership_type: String,
-    pub ownership_type_2: String,
-    pub previous_assessed_value: String,
-    pub prior_sale_amount: String,
-    pub prior_sale_date: String,
-    pub sale_amount: String,
-    pub sale_date: String,
-    pub senior_tax_exemption: String,
-    pub suffix: String,
-    pub suffix_2: String,
-    pub suffix_3: String,
-    pub suffix_4: String,
-    pub tax_assess_year: String,
-    pub tax_billed_amount: String,
-    pub tax_delinquent_year: String,
-    pub tax_fiscal_year: String,
-    pub tax_rate_area: String,
-    pub total_market_value: String,
-    pub trust_description: String,
-    pub veteran_tax_exemption: String,
-    pub widow_tax_exemption: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(default)]
-pub struct FinancialHistory {
-    pub code_title_company: String,
-    pub document_type_description: String,
-    pub instrument_date: String,
-    pub interest_rate_type_2: String,
-    pub lender_address: String,
-    pub lender_address_2: String,
-    pub lender_city: String,
-    pub lender_city_2: String,
-    pub lender_code_2: String,
-    pub lender_first_name: String,
-    pub lender_first_name_2: String,
-    pub lender_last_name: String,
-    pub lender_last_name_2: String,
-    pub lender_name: String,
-    pub lender_name_2: String,
-    pub lender_seller_carry_back: String,
-    pub lender_seller_carry_back_2: String,
-    pub lender_state: String,
-    pub lender_state_2: String,
-    pub lender_zip: String,
-    pub lender_zip_2: String,
-    pub lender_zip_extended: String,
-    pub lender_zip_extended_2: String,
-    pub mortgage_amount: String,
-    pub mortgage_amount_2: String,
-    pub mortgage_due_date: String,
-    pub mortgage_due_date_2: String,
-    pub mortgage_interest_rate: String,
-    pub mortgage_interest_rate_type: String,
-    pub mortgage_lender_code: String,
-    pub mortgage_rate_2: String,
-    pub mortgage_recording_date: String,
-    pub mortgage_recording_date_2: String,
-    pub mortgage_term: String,
-    pub mortgage_term_2: String,
-    pub mortgage_term_type: String,
-    pub mortgage_term_type_2: String,
-    pub mortgage_type: String,
-    pub mortgage_type_2: String,
-    pub multi_parcel_flag: String,
-    pub name_title_company: String,
-    pub recording_date: String,
-    pub transfer_amount: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct FinancialResponse {
-    pub smarty_key: String,
-    pub data_set_name: String,
-    pub data_subset_name: String,
-    pub attributes: FinancialAttributes,
-}
-
-impl EnrichmentResponse for FinancialResponse {
-    fn lookup_type() -> &'static str {
-        "financial"
-    }
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct GeoReferenceResponse {
-    pub census_block: CensusBlock,
-    pub census_county_division: CensusCountyDivision,
-    pub census_tract: CensusTract,
-    pub core_based_stat_area: CoreBasedStatArea,
-    pub place: Place,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct CensusBlock {
-    pub accuracy: CensusGeoIdAccuracy,
-    pub geoid: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum CensusGeoIdAccuracy {
-    #[default]
-    #[serde(rename = "block")]
-    Block,
-    #[serde(rename = "tract")]
-    Tract,
-    #[serde(rename = "county")]
-    County,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum GeocodeAccuracy {
-    #[default]
-    #[serde(rename = "inferred")]
-    Inferred,
-    #[serde(rename = "exact")]
-    Exact,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct CensusCountyDivision {
-    accuracy: GeocodeAccuracy,
-    code: String,
-    name: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct CensusTract {
-    code: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct CoreBasedStatArea {
-    code: String,
-    name: String,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct Place {
-    accuracy: GeocodeAccuracy,
-    code: String,
-    name: String,
-    #[serde(rename = "type")]
-    type_: PlaceType,
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub enum PlaceType {
-    #[default]
-    #[serde(rename = "unincorporated")]
-    Unincorperated,
-    #[serde(rename = "incorperated")]
-    Incorperated,
-}
-
-impl EnrichmentResponse for GeoReferenceResponse {
-    fn lookup_type() -> &'static str {
-        "geo-reference"
-    }
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct GeoReference2010Response(pub GeoReferenceResponse);
-
-impl Deref for GeoReference2010Response {
-    type Target = GeoReferenceResponse;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl DerefMut for GeoReference2010Response {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-impl EnrichmentResponse for GeoReference2010Response {
-    fn lookup_type() -> &'static str {
-        "geo-reference/2010"
-    }
-}
-
-#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct GeoReference2020Response(pub GeoReferenceResponse);
-
-impl Deref for GeoReference2020Response {
-    type Target = GeoReferenceResponse;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl DerefMut for GeoReference2020Response {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-impl EnrichmentResponse for GeoReference2020Response {
-    fn lookup_type() -> &'static str {
-        "geo-reference/2020"
-    }
 }
