+++ smarty-rust-sdk/src/us_enrichment_api/response.rs
@@ -1,0 +1,6 @@
+use serde::de::DeserializeOwned;
+use serde::Serialize;
+
+pub trait EnrichmentResponse: Clone + Serialize + DeserializeOwned + Default {
+    fn lookup_type() -> &'static str;
+}
